배송지 등록,수정,삭제시 validation 처리
이제 배송지를 등록할 때 입력 validation이 동작합니다
예외는 IllegalArgumentExcpetion을 impl한
AddressIllegalArgumentException으로 발생하고
config 폴더의 RestExceptionHandler에서 처리됩니다

+++ src/main/java/com/nhnacademy/taskapi/address/controller/AddressController.java
@@ -4,10 +4,13 @@ import com.nhnacademy.taskapi.address.domain.dto.req.AddMemberAddressRequest;
 import com.nhnacademy.taskapi.address.domain.dto.req.DeleteMemberAddressRequest;
 import com.nhnacademy.taskapi.address.domain.dto.req.UpdateMemberAddressRequest;
 import com.nhnacademy.taskapi.address.domain.dto.resp.MemberAddressResponse;
+import com.nhnacademy.taskapi.address.exception.AddressIllegalArgumentException;
 import com.nhnacademy.taskapi.address.service.AddressService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,7 +24,12 @@ public class AddressController {
 
     @PostMapping("/addresses")
     public ResponseEntity<MemberAddressResponse> addMemberAddress(@RequestHeader("X-MEMBER-ID") Long memberId ,
-                                                                  @RequestBody AddMemberAddressRequest request){
+                                                                  @Valid @RequestBody AddMemberAddressRequest request,
+                                                                  BindingResult bindingResult){
+
+        if(bindingResult.hasErrors()){
+            throw new AddressIllegalArgumentException(bindingResult.getAllErrors().getFirst().getDefaultMessage());
+        }
 
         MemberAddressResponse resp = addressService.addMemberAddress(memberId, request);
         return new ResponseEntity<>(resp,HttpStatus.OK);
@@ -42,14 +50,23 @@ public class AddressController {
     }
     
     @PutMapping("/addresses")
-    public ResponseEntity<MemberAddressResponse> updateMemberAddress(@RequestHeader("X-MEMBER-ID") Long memberId, @RequestBody UpdateMemberAddressRequest request){
+    public ResponseEntity<MemberAddressResponse> updateMemberAddress(@RequestHeader("X-MEMBER-ID") Long memberId
+            ,@Valid @RequestBody UpdateMemberAddressRequest request, BindingResult bindingResult){
 
+        if(bindingResult.hasErrors()){
+            throw new AddressIllegalArgumentException(bindingResult.getAllErrors().getFirst().getDefaultMessage());
+        }
         MemberAddressResponse resp = addressService.updateMemberAddress(memberId, request);
         return new ResponseEntity<>(resp,HttpStatus.OK);
     }
 
     @DeleteMapping("/addresses")
-    public ResponseEntity<MemberAddressResponse> deleteMemberAddress(@RequestHeader("X-MEMBER-ID") Long memberId, @RequestBody DeleteMemberAddressRequest request){
+    public ResponseEntity<MemberAddressResponse> deleteMemberAddress(@RequestHeader("X-MEMBER-ID") Long memberId
+            ,@Valid @RequestBody DeleteMemberAddressRequest request, BindingResult bindingResult){
+
+        if(bindingResult.hasErrors()){
+            throw new AddressIllegalArgumentException(bindingResult.getAllErrors().getFirst().getDefaultMessage());
+        }
 
         MemberAddressResponse resp = addressService.deleteMemberAddress(memberId,request);
         return new ResponseEntity<>(resp,HttpStatus.OK);
